% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_MLCL.R
\name{create_MLCL}
\alias{create_MLCL}
\title{Random generation of Must-Link and Cannot-Link constraints}
\usage{
create_MLCL(y, nbConst)
}
\arguments{
\item{y}{Vector of class labels.}

\item{nbConst}{Number of constraints.}
}
\value{
A list with two components:
\describe{
  \item{ML}{Matrix of ML constraints. Each row corresponds to a constraint.}
  \item{CL}{Matrix of ML constraints. Each row corresponds to a constraint.}
}
}
\description{
\code{create_MLCL} randomly generates Must-Link (ML) and Cannot-Link (CL) constraints from a vector y
of class labels.
}
\examples{
y<-sample(3,100,replace=TRUE)
const<-create_MLCL(y,nbConst=10)
const$ML
const$CL

}
\seealso{
\code{\link{cecm}}
}
