% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourclass-data.R
\docType{data}
\name{fourclass}
\alias{fourclass}
\title{Synthetic four-class dataset}
\format{
A data frame with three variables: x1, x2 and y (the true class).
}
\usage{
data(fourclass)
}
\description{
A synthetic dataset with two attributes and four classes of 100 points each,  generated from a
multivariate t distribution with five degrees of freedom and centered, respectively,
on [0;0], [0;4], [4;0] and [4;4].
}
\examples{
data(fourclass)
plot(fourclass$x1,fourclass$x2,xlab=expression(x[1]),ylab=expression(x[2]),
col=fourclass$y,pch=fourclass$y)
}
\references{
M.-H. Masson and T. Denoeux. ECM: An evidential version of the fuzzy c-means algorithm.
Pattern Recognition, Vol. 41, Issue 4, pages 1384-1397, 2008.
}
\keyword{datasets}
